More Practice questions done

+++ C programmes/Practice/Larger.c
@@ -1,22 +1,22 @@
 #include <stdio.h>
 int main()
 {
-    int a,b;
+    float a,b;
 
     printf("Enter first number: ");
-    scanf("%d", &a);
+    scanf("%f", &a);
 
     printf("\nEnter second number: ");
-    scanf("%d", &b);
+    scanf("%f", &b);
 
     if(a>b)
-    {printf("%d is larger", a);}
+    {printf("%f is larger", a);}
 
     else if(a == b)
     {printf("Both numbers are equal");}
 
     else
-    {printf("%d is larger", b);}
+    {printf("%f is larger", b);}
 
     return 0;
